% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_detail_old.R
\name{coxph_mi}
\alias{coxph_mi}
\title{Multiple Imputation with Rubin's rule and Wild Bootstrap}
\usage{
coxph_mi(
  fit,
  pattern,
  delta,
  n_mi,
  n_wb,
  cut_point,
  seed = NULL,
  validate = FALSE
)
}
\arguments{
\item{fit}{a Cox model object, i.e., the result of \code{coxph}.}

\item{pattern}{A integer vector of pattern indicator for each subject
* 1  = observed event time
* 2  = censoring at random
* 3+ = censoring not at random}

\item{delta}{A numeric vector of adjusted delta in the same length of pattern.
* delta = 1 is for observed event time and censoring at random.
* delta > 1 is for censoring time with hazard is larger than the hazard in CAR situation.
* delta < 1 is for censoring time with hazard is smaller than the hazard in CAR situation.}

\item{n_mi}{Number of multiple imputation}

\item{n_wb}{Number of bootstrap replications}

\item{cut_point}{Cutpoint of RMST}

\item{seed}{Randomization seed}

\item{validate}{Validation model, default is false. Only used for testing purpose.}
}
\description{
Multiple Imputation with Rubin's rule and Wild Bootstrap
}
