% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_martingale.R
\name{km_martingale}
\alias{km_martingale}
\title{Martingale Details of a Kaplan Meier Estmator}
\usage{
km_martingale(fit)
}
\arguments{
\item{fit}{a Cox model object, i.e., the result of \code{coxph}.}
}
\description{
Martingale Details of a Kaplan Meier Estmator
}
