% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmst.R
\name{rmst_control}
\alias{rmst_control}
\title{Control-based imputation for RMST}
\usage{
rmst_control(
  time,
  status,
  x,
  group,
  ref_grp = 0,
  pattern,
  delta,
  tau,
  n_mi = 10,
  n_b = 100,
  seed = NULL,
  wild_boot = TRUE,
  validate = FALSE
)
}
\description{
Control-based imputation for RMST
}
