% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmst.R
\name{rmst_delta}
\alias{rmst_delta}
\title{Delta adjusted RMST within group}
\usage{
rmst_delta(
  time,
  status,
  x,
  group,
  pattern,
  delta,
  tau,
  n_mi = 10,
  n_b = 100,
  seed = NULL,
  wild_boot = TRUE,
  validate = FALSE
)
}
\description{
Delta adjusted RMST within group
}
